requirements.txt
pyyaml
pymemgpt